move SingleXAResource to SPI for other transaction manager integration.
