+++ sharding-transaction/sharding-transaction-2pc/sharding-transaction-xa/sharding-transaction-xa-core/src/test/java/io/shardingsphere/transaction/xa/jta/AllJTATests.java
@@ -31,15 +31,13 @@ import io.shardingsphere.transaction.xa.jta.datasource.properties.dialect.SQLSer
 import io.shardingsphere.transaction.xa.jta.datasource.swapper.DataSourcePropertyProviderLoaderTest;
 import io.shardingsphere.transaction.xa.jta.datasource.swapper.DataSourceSwapperTest;
 import io.shardingsphere.transaction.xa.jta.datasource.swapper.impl.DefaultDataSourcePropertyProviderTest;
-import io.shardingsphere.transaction.xa.jta.resource.SingleXAResourceTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
 @SuiteClasses({
-        SingleXAResourceTest.class,
-        XAPropertiesFactoryTest.class, 
+        XAPropertiesFactoryTest.class,
         SingleXADataSourceTest.class,
         XADataSourceFactoryTest.class, 
         DataSourcePropertyProviderLoaderTest.class,
