+++ sharding-transaction/sharding-transaction-2pc/sharding-transaction-xa/sharding-transaction-xa-core/src/test/java/io/shardingsphere/transaction/xa/manager/atomikos/AtomikosXARecoverableResourceTest.java
@@ -17,7 +17,7 @@
 
 package io.shardingsphere.transaction.xa.manager.atomikos;
 
-import io.shardingsphere.transaction.xa.jta.resource.SingleXAResource;
+import io.shardingsphere.transaction.xa.spi.SingleXAResource;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
