+++ sharding-transaction/sharding-transaction-2pc/sharding-transaction-xa/sharding-transaction-xa-spi/src/main/java/io/shardingsphere/transaction/xa/spi/SingleXAResource.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.jta.resource;
+package io.shardingsphere.transaction.xa.spi;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
