+++ sharding-transaction/sharding-transaction-2pc/sharding-transaction-xa/sharding-transaction-xa-spi/src/test/java/io/shardingsphere/transaction/xa/spi/SingleXAResourceTest.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.jta.resource;
+package io.shardingsphere.transaction.xa.spi;
 
 import org.junit.Before;
 import org.junit.Test;
